Restructure docs files
