+++ docs/src/dashboard/DashboardLayout.jsx
